// src/types/storeTypes.tsx
export interface INavigationData {
  list: string;
  to: string;
}

export interface IOfferBannersData {
  img: string;
  colorTitle: string;
  title: string;
  colorText: string;
  text: string;
}

export interface IAboutData {
  icon: string;
  textTitle: string;
  text: string;
}

export interface IReviewsData {
  title: string;
  text: string;
}

export interface IOffersData {
  type: string;
  photo: string;
  name: string;
}

export interface IWhoWeAreData {
  title: string;
  text: string;
}

export interface IGalleryData {
  backgroundImage: string;
  btnText: string;
}

export interface INewsData {
  backgroundImage: string;
  date: string;
  title: string;
  supTitle: string;
}

export interface IFooterContactsData {
  title: string;
  subtitle: string;
}

export interface ISocialListData {
  backgroundImage: string;
  altText: string;
}

export interface IBlogData {
  backgroundImage: string;
  date: string;
  title: string;
  supTitle: string;
}